Adds websockets with redis messaging for real time moves

+++ backend/src/testUtils/db/handler.ts
@@ -11,7 +11,9 @@ export const dbConnect = async () => {
 }
 
 export const dbDisconnect = async () => {
-  await mongoose.connection.dropDatabase()
-  await mongoose.connection.close()
+  if (mongoose.connection.readyState === 1) {
+    await mongoose.connection.dropDatabase()
+    await mongoose.connection.close()
+  }
   await mongoServer.stop()
 }
